backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
bcrypt==4.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
requests==2.31.0
python-dateutil==2.8.2

# ML Libraries
joblib==1.4.2
scikit-learn==1.7.2
pandas==2.2.3
numpy==2.1.3
xgboost==2.1.1
lightgbm==4.5.0
scipy==1.14.1

# Production server
gunicorn==21.2.0
psycopg2-binary==2.9.9

# PDF and LLM libraries
google-generativeai==0.3.2
fpdf2==2.7.6

